Fixed defect (PLATYPUS-39) in which a file consisting solely of a code listing did not generate output. Modified PdfCodeOff.java so that if the code is a listing, when it hits this command the listing content is then and there written to the output file.
This required changing PdfOutfile so that when a paragraph is added to the output PDF file, it checks to see whether the file is open and it takes the necessary steps if not.

Also, added a flag to PdfData (actually, to its superclass, DocData) so that if a listing is being generated (via code with options command), then [-code] can detect this and act accordingly. 

Wrote a functional test for this in pdfOperation.groovy