PLATYPUS-61 Changed PdfBulletListPlainEnd.java to take into account the fact that the iText library adds a CR/LF when the outermost unordered list ends.
So, when it's followed by a new paragraph ([CR]) an extra blank line is generated. This is solved in the above changes by temporarily decreasing the number of blank lines between paragraphs and then restoring it once the new paragraph is created.

To do: Still needs unit testing and any updates to the documentation.