PLATYPUS-66 Fixed problem with UrlWithCoverText.java superclassIt was balking at this code, which checks whether the command the URL text ends with a [-url] command:
if( tok.getType() == TokenType.COMMAND && tok.getRoot().equals( new PdfUrlWithCoverTextEnd().getRoot() )) 

The creation of the new class was causing a NoClassDefFound exception. 

Replaced the second half of the if statement with tok.getRoot().equals("[-url]") This solves the problem at the cost of hard-coding the command into this class, alas.